+++ lib/src/win32/wlanapi.g.dart
@@ -1134,7 +1134,7 @@ int WlanRegisterNotification(
         int hClientHandle,
         int dwNotifSource,
         int bIgnoreDuplicate,
-        Pointer<NativeFunction<WlanNotificationCallback>> funcCallback,
+        Pointer<NativeFunction<WLAN_NOTIFICATION_CALLBACK>> funcCallback,
         Pointer pCallbackContext,
         Pointer pReserved,
         Pointer<Uint32> pdwPrevNotifSource) =>
@@ -1146,7 +1146,7 @@ final _WlanRegisterNotification = _wlanapi.lookupFunction<
         IntPtr hClientHandle,
         Uint32 dwNotifSource,
         Int32 bIgnoreDuplicate,
-        Pointer<NativeFunction<WlanNotificationCallback>> funcCallback,
+        Pointer<NativeFunction<WLAN_NOTIFICATION_CALLBACK>> funcCallback,
         Pointer pCallbackContext,
         Pointer pReserved,
         Pointer<Uint32> pdwPrevNotifSource),
@@ -1154,7 +1154,7 @@ final _WlanRegisterNotification = _wlanapi.lookupFunction<
         int hClientHandle,
         int dwNotifSource,
         int bIgnoreDuplicate,
-        Pointer<NativeFunction<WlanNotificationCallback>> funcCallback,
+        Pointer<NativeFunction<WLAN_NOTIFICATION_CALLBACK>> funcCallback,
         Pointer pCallbackContext,
         Pointer pReserved,
         Pointer<Uint32> pdwPrevNotifSource)>('WlanRegisterNotification');